backend/requirements.txt
django==5.0.6
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0
scikit-learn>=1.4.0
langdetect>=1.0.9
textblob>=0.17.1
vaderSentiment>=3.3.2
requests>=2.31.0
bleach>=6.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
social-auth-app-django==5.4.1
langid>=1.1.6

